moved any helper funcs into utils.go
